Ease of use changes
Added a function for my code to be easy to change.

Added the test videos

Implemented Renaud's finidng circle position algo and made my funciton write the postions in a text file and return them

Added a main file where all the long function will be imported and the code will hence be easier to read.